# kvgraph: further performance improvements