--- a/plugins/kvscheduler/internal/graph/bench_test.go
+++ b/plugins/kvscheduler/internal/graph/bench_test.go
@@ -53,7 +53,7 @@ type scaleCtx struct {
 }
 
 func init() {
-	benchEl = newEdgeLookup()
+	benchEl = newEdgeLookup(nil)
 }
 
 func BenchmarkScaleWithoutRec(b *testing.B) {