Improved `edgeLookup` from `kvscheduler/internal/graph` to avoid very expensive memmove of data with pointers - from the CPU profile of GRPC perf test:
```
      flat  flat%   sum%        cum   cum%
    12.83s 33.46% 33.46%     12.83s 33.46%  runtime.memmove
     3.20s  8.34% 41.80%      3.20s  8.34%  runtime.futex
     2.59s  6.75% 48.55%      2.92s  7.61%  syscall.Syscall
     1.41s  3.68% 52.23%      1.41s  3.68%  runtime.usleep
     0.74s  1.93% 54.16%      0.74s  1.93%  runtime.epollwait
     0.68s  1.77% 55.93%      1.57s  4.09%  runtime.scanobject
     0.52s  1.36% 57.29%      2.19s  5.71%  runtime.mallocgc
     0.44s  1.15% 58.44%      0.44s  1.15%  runtime.memclrNoHeapPointers
     0.41s  1.07% 59.50%      0.41s  1.07%  cmpbody
     0.35s  0.91% 60.42%      0.42s  1.10%  runtime.findObject
```

With this improvement, the time spent in `Graph.SetNode` and especially in `Graph.SetTargets` should decrease significantly. 

Signed-off-by: Milan Lenco <milan.lenco@pantheon.tech>